requirements.txt
streamlit
pandas
python-docx
requests
msal
openai
transformers
torch
Pillow
PyMuPDF